requirements.txt
streamlit>=1.30.0
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-dotenv==1.0.1
streamlit>=1.30.0
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-dotenv==1.0.1
